feat: rework the way Sidero Agent boots and configures networking
1. Update `PKGS` to the latest version which has kernel with modules.
2. Pull in `udevd`, kernel modules into the Sidero Agent initramfs.
3. Stop doing kernel network configuration (`ip=dhcp`).
4. Run `udevd` in `initramfs` to auto-load kernel modules based on
   hardware information.
5. Pass MAC address of the NIC used to PXE boot from iPXE down to Sidero
   Agent.
6. In the Sidero Agent, run DHCP configuration on the link passed by
   MAC.

Slightly refactored Sidero Agent code by pulling apart `main.go` into
smaller pieces.
